+++ js/streams/client/event.ts
@@ -4,6 +4,7 @@ export interface SubscriptionEvent extends BaseEvent {
     topic: string;
     raisedAt: Date;
     orderSerial?: number;
+    deploymentId?: number;
 }
 
 export interface PublishEvent extends BaseEvent {
@@ -20,7 +21,6 @@ export interface BaseEvent {
     causationId?: string;
     reason?: string;
     userId?: string;
-    deploymentId?: number;
 }
 
 /* eslint-disable-next-line @typescript-eslint/no-explicit-any */